Add feature: deleting posts
Deleting a post won't actually delete it, it will replace it with a system post (new type of post) with the content of "Post deleted by user" and no author. This type of post will not show up on post lists (such as those in profile, search and home page).